Move account-nft tests.
