+++ contracts/account-nft/tests/tests/mod.rs
@@ -1,0 +1,9 @@
+mod helpers;
+
+mod test_burn_allowance;
+mod test_burn_empty_accounts;
+mod test_instantiate;
+mod test_migration;
+mod test_mint;
+mod test_proposed_minter;
+mod test_update_config;
